2. Тесты для идеальный вес[#266407] исправил тип параметра in.

+++ test/ru/job4j/calculator/FitTest.java
@@ -7,7 +7,7 @@ public class FitTest {
 
     @Test
     public void whenManHeight180Then92Weight() {
-        int in = 180;
+        double in = 180;
         double expected = 92;
         double out = Fit.manWeight(in);
         Assert.assertEquals(expected, out, 0.01);
@@ -15,7 +15,7 @@ public class FitTest {
 
     @Test
     public void whenWomanHeight170Then69Weight() {
-        int in = 170;
+        double in = 170;
         double expected = 69;
         double out = Fit.womanWeight(in);
         Assert.assertEquals(expected, out, 0.01);
